requirements.txt
appdirs==1.4.4
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cffi==1.14.4
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
flake8==3.8.4
Flask==1.1.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
parsel==1.6.0
Protego==0.1.16
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
pyee==8.1.0
pyflakes==2.2.0
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyppeteer==0.2.5
queuelib==1.5.0
Scrapy==2.4.1
scrapy-selenium==0.0.7
scrapy-splash==0.7.2
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
tqdm==4.56.0
Twisted==20.3.0
urllib3==1.26.2
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1
zope.interface==5.2.0
